Add files via upload
Session Hijacking Fixation is using two browsers but you are logged in on one and not the other but you use PHPSESSIONID to change it to the same one on the browser that is not logged in. And once you change it to the same you go to the one that is not logged in and it will automatically log you in.